backend/requirements.txt
Django==3.2.18
djoser==2.1.0
Pillow==9.5.0
django-filter==23.1
django-extra-fields==3.0.2
psycopg2-binary==2.9.6
